Moves ferries packages into hierarchical nav order
