+++ app/src/main/java/gov/wa/wsdot/android/wsdot/ui/ferries/schedules/bulletins/FerriesRouteAlertsBulletinDetailsActivity.java
@@ -16,7 +16,7 @@
  *
  */
 
-package gov.wa.wsdot.android.wsdot.ui.ferries.bulletins;
+package gov.wa.wsdot.android.wsdot.ui.ferries.schedules.bulletins;
 
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
