+++ app/src/main/java/gov/wa/wsdot/android/wsdot/ui/ferries/schedules/sailings/FerriesRouteSchedulesDaySailingsActivity.java
@@ -16,16 +16,14 @@
  *
  */
 
-package gov.wa.wsdot.android.wsdot.ui.ferries.sailings;
+package gov.wa.wsdot.android.wsdot.ui.ferries.schedules.sailings;
 
 import android.arch.lifecycle.ViewModelProvider;
 import android.arch.lifecycle.ViewModelProviders;
-import android.content.ContentResolver;
 import android.os.Bundle;
 import android.support.design.widget.Snackbar;
 import android.support.v4.view.MenuItemCompat;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.accessibility.AccessibilityEvent;
