+++ app/src/main/res/layout/activity_ferries_route_alerts_bulletins.xml
@@ -7,7 +7,7 @@
     <include layout="@layout/toolbar_actionbar"/>
 
     <fragment
-        android:name="gov.wa.wsdot.android.wsdot.ui.ferries.bulletins.FerriesRouteAlertsBulletinsFragment"
+        android:name="gov.wa.wsdot.android.wsdot.ui.ferries.schedules.bulletins.FerriesRouteAlertsBulletinsFragment"
         android:id="@+id/fragment_ferries_route_alerts_bulletins"
         android:layout_width="match_parent"
         android:layout_height="0dp"
