+++ app/src/main/res/layout/activity_ferries_route_schedules_day_sailings.xml
@@ -15,7 +15,7 @@
         <include layout="@layout/toolbar_scrolling"/>
 
         <fragment
-            android:name="gov.wa.wsdot.android.wsdot.ui.ferries.sailings.FerriesRouteSchedulesDaySailingsFragment"
+            android:name="gov.wa.wsdot.android.wsdot.ui.ferries.schedules.sailings.FerriesRouteSchedulesDaySailingsFragment"
             android:id="@+id/fragment_ferries_route_schedules_day_sailings"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
